requirements.txt
python-dotenv
moviepy
SpeechRecognition
